Customer Deatils Issue

+++ app/Http/Controllers/ApiController.php
@@ -133,7 +133,9 @@ class ApiController extends Controller
         $customerInfo = $request->all();
         unset($customerInfo['amount']);
         $customer = User::with('CustomerDetails')->find($request->id);
-        $customer->CustomerDetails()->delete();
+        if(!is_null($customer->CustomerDetails())) {
+            $customer->CustomerDetails()->delete();
+        }
         $customer->CustomerDetails()->create($customerInfo);
 
         $api = new Api(PaymentApiConfig()->payKeyId,PaymentApiConfig()->paySecretKey);
